Create create_facility_list_item mettod for each class ProcessingFacilityList and ProcessingFacilityAPI
