+++ src/django/api/facility_actions/processing_facility.py
@@ -25,23 +25,23 @@ class ProcessingFacility(ABC):
     def process_facility(self):
         pass
 
-    @staticmethod
-    def _create_facility_list_item(
-        source: Source, row: RowDTO, idx: int, header_str: str
-    ) -> FacilityListItem:
-        return FacilityListItem.objects.create(
-            source=source,
-            row_index=idx,
-            raw_data=','.join(f'"{value}"' for value in row.raw_json.values()),
-            raw_json=row.raw_json,
-            raw_header=header_str,
-            name=row.name,
-            clean_name=row.clean_name,
-            address=row.address,
-            clean_address=row.clean_address,
-            country_code=row.country_code,
-            sector=row.sector,
-        )
+    # @staticmethod
+    # def _create_facility_list_item(
+    #     source: Source, row: RowDTO, idx: int, header_str: str
+    # ) -> FacilityListItem:
+    #     return FacilityListItem.objects.create(
+    #         source=source,
+    #         row_index=idx,
+    #         raw_data=','.join(f'"{value}"' for value in row.raw_json.values()),
+    #         raw_json=row.raw_json,
+    #         raw_header=header_str,
+    #         name=row.name,
+    #         clean_name=row.clean_name,
+    #         address=row.address,
+    #         clean_address=row.clean_address,
+    #         country_code=row.country_code,
+    #         sector=row.sector,
+    #     )
 
     @staticmethod
     def _create_nonstandard_fields(
