+++ src/django/api/facility_actions/processing_facility_api.py
@@ -12,6 +12,7 @@ from api.models.facility.facility_list_item import FacilityListItem
 from api.models.facility.facility_list_item_temp import FacilityListItemTemp
 from api.models.source import Source
 from api.processing import handle_external_match_process_result
+from contricleaner.lib.dto.row_dto import RowDTO
 from rest_framework import status
 from rest_framework.response import Response
 
@@ -255,4 +256,22 @@ class ProcessingFacilityAPI(ProcessingFacility):
             source_type=Source.SINGLE,
             is_public=public_submission,
             create=should_create,
+        )
+
+    @staticmethod
+    def _create_facility_list_item(
+        source: Source, row: RowDTO, idx: int, header_str: str
+    ) -> FacilityListItem:
+        return FacilityListItem.objects.create(
+            source=source,
+            row_index=idx,
+            raw_data=','.join(f'"{value}"' for value in row.raw_json.values()),
+            raw_json=row.raw_json,
+            raw_header=header_str,
+            name=row.name,
+            clean_name=row.clean_name,
+            address=row.address,
+            clean_address=row.clean_address,
+            country_code=row.country_code,
+            sector=row.sector,
         )
